pyproject.toml
# pyproject.toml
[project]
name = "fastapi-core-module"
version = "1.0.0"
description = "Основной модуль для FastAPI приложений с PostgreSQL и Redis"
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.2.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]